Persist the username on firebase
There was an issue where the wrong username was being saved. This was due to the previous user still being logged in (no logout function at the moment). I now create a user, log them in and then only assign the display name to the newly created and logged in user